Move actions module to top-level
