+++ crates/modalkit-ratatui/examples/editor.rs
@@ -23,28 +23,22 @@ use ratatui::{
 };
 
 use modalkit::{
+    actions::{
+        Action,
+        Commandable,
+        Editable,
+        EditorAction,
+        Jumpable,
+        PromptAction,
+        Promptable,
+        Scrollable,
+        Searchable,
+        TabContainer,
+        TabCount,
+        WindowAction,
+        WindowContainer,
+    },
     editing::{
-        action::{
-            Action,
-            Commandable,
-            EditError,
-            EditInfo,
-            EditResult,
-            Editable,
-            EditorAction,
-            InfoMessage,
-            Jumpable,
-            PromptAction,
-            Promptable,
-            Scrollable,
-            Searchable,
-            TabContainer,
-            TabCount,
-            UIError,
-            UIResult,
-            WindowAction,
-            WindowContainer,
-        },
         application::{
             ApplicationContentId,
             ApplicationInfo,
@@ -61,6 +55,7 @@ use modalkit::{
     },
     env::mixed::{MixedBindings, MixedChoice},
     env::vim::command::{complete_cmdbar, VimCommandMachine},
+    errors::{EditError, EditResult, UIError, UIResult},
     key::TerminalKey,
     keybindings::{dialog::Pager, BindingMachine},
     prelude::*,
